# laravel-bek/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Patreon Clone API",
        "description": "API dokumentacija za platformu za podršku kreatorima",
        "contact": {
            "name": "API Support",
            "url": "https://example.com/support",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/creators/{id}/subscribe": {
            "post": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Subscribe to a creator",
                "description": "Subscribe to a creator.",
                "operationId": "c8e2879095610435454d5d2a38d56dab",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Creator ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nivo_id": {
                                        "description": "Tier ID (optional)",
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Subscribed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubscriptionResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Cannot subscribe to yourself"
                    },
                    "404": {
                        "description": "Creator not found"
                    },
                    "409": {
                        "description": "Already subscribed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Unsubscribe from a creator",
                "description": "Unsubscribe from a creator.",
                "operationId": "cea31e406cbd13d7518c63b4896b576e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Creator ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Unsubscribed"
                    },
                    "404": {
                        "description": "Not subscribed or creator not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/subscriptions": {
            "get": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "List authenticated user's subscriptions",
                "description": "List authenticated user's subscriptions.",
                "operationId": "67df8b16e2d36e7aab97ce19a17f9355",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of subscriptions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SubscriptionResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/subscriptions/{id}": {
            "get": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Show a single subscription",
                "description": "Show a single subscription (only if it belongs to the user).",
                "operationId": "6b9cf0d3971878009b6d4f0a5598dae9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subscription details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubscriptionResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Update subscription tier",
                "description": "Update subscription tier.",
                "operationId": "bac8be53d2f4233f44fd60abb494230b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nivo_id": {
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tier updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubscriptionResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Subscription not found"
                    },
                    "422": {
                        "description": "Tier does not belong to creator"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/my-subscriptions/total-cost": {
            "get": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Vrati mesecni trosak svih aktivnih pretplata",
                "description": "Vraca ukupni mesecni trosak pretplata od ulogovanog korisnika tipa patron",
                "operationId": "954d502b0ae3c8dbaefaa7b7c8d209fe",
                "responses": {
                    "200": {
                        "description": "Ukupni mesecni trosak pretplata i lista pretplata",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "subscriptions": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "creator_name": {
                                                        "type": "string"
                                                    },
                                                    "tier_name": {
                                                        "type": "string"
                                                    },
                                                    "price": {
                                                        "type": "number",
                                                        "format": "float"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "total_cost": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 29.97
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SubscriptionResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 5
                    },
                    "creator": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 2
                            },
                            "naziv_stranice": {
                                "type": "string",
                                "example": "Muzički kanal"
                            },
                            "user": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Marko Marković"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object"
                    },
                    "tier": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 3
                            },
                            "naziv": {
                                "type": "string",
                                "example": "Gold"
                            },
                            "cena_mesecno": {
                                "type": "number",
                                "format": "float",
                                "example": 9.99
                            }
                        },
                        "type": "object",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "aktivna",
                        "enum": [
                            "aktivna",
                            "otkazana",
                            "istekla"
                        ]
                    },
                    "datum_pocetka": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "sanctum",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}